Hoist needless function-level imports in the strategy-system modules
Function-level imports exist for two reasons — CUDA/heavy optional deps
(cupy, catboost) and real import cycles — and many locals in the modules
this PR created or rewrote guarded neither (stdlib, pure IR types, modules
the file already imported at top). Hoisted throughout terminal_bench,
two_level, strategy/greedy, identity, candidate, mcts, and policy/base.
O3_NVCC_FLAGS moves from mcts to terminal_bench (it documents that
module's re-bench), turning the mcts→terminal_bench edge into a clean
top-level import instead of a lazy one dodging a cycle.

What stays local is exactly the justified set: cupy (terminal_bench's
memory-pool cleanup), the prior's catboost-backed load, mcts's documented
golden-layering import, and pipeline.py's lazy GreedyStrategy (a genuine
package-init cycle).
