+++ emmy/compiler/pipeline/search/policy/base.py
@@ -28,6 +28,7 @@ from abc import ABC, abstractmethod
 
 from emmy.compiler.pipeline.search.candidate import LazyCandidate
 from emmy.compiler.pipeline.search.db import PerfStats
+from emmy.compiler.pipeline.search.policy.terminal_bench import bench_terminal_async
 
 
 class Search(ABC):
@@ -87,8 +88,6 @@ class Search(ABC):
         engine only awaits it. The default benches the terminal's kernels (cache/stub-aware) and
         feeds :meth:`observe`; :class:`~.mcts.TuningSearch` extends it with the deployable -O3
         re-bench and the prior's row protocol."""
-        from emmy.compiler.pipeline.search.policy.terminal_bench import bench_terminal_async  # noqa: PLC0415
-
         stats, status, measured, per_kernel = await bench_terminal_async(cand, backend=backend, db=db)
         self.note_bench(measured=measured)
         self.observe(token, stats, status, candidate=cand, kernels=per_kernel)
